src/app/core/validators/inscricao-estadual-estados/amapa.validator.spec.ts
import { Amapa } from './amapa.validator';
import { InscricaoEstadualFactory } from './inscricao-estadual.factory';

describe('classe validacao: Inscricao Estadual Amapa', () => {

  const estado = InscricaoEstadualFactory.criaInstancia(InscricaoEstadualFactory.AMAPA);

  it('instancia deve estar criada', () => {
    expect(estado).toBeTruthy();
  });

  it('inscricao estadual deve estar valida', () => {
    expect(estado.validar('033442959')).toBe(true);
    expect(estado.validar('030329086')).toBe(true);
    expect(estado.validar('039401243')).toBe(true);
    expect(estado.validar('032421540')).toBe(true);
  });

  it('inscricao estadual deve estar invalida', () => {
    expect(estado.validar('123461')).toBe(false);
    expect(estado.validar('31632')).toBe(false);
    expect(estado.validar('213413513461346')).toBe(false);
  });

  it('inscricao estadual deve estar invalida quando valor for nulo', () => {
    expect(estado.validar(null)).toBe(false);
  });

  it('inscricao estadual deve estar invalida quando valor for undefined', () => {
    expect(estado.validar(undefined)).toBe(false);
  });

});

